App2/requirements.txt
python-dotenv
databricks-sdk
fastapi
uvicorn
jinja2
psycopg2-binary
